Move socket enricher to pkg/
Move it to a non-internal pkg to use it on the next commits.
