+++ gadgets/trace_dns/program.bpf.c
@@ -254,7 +254,7 @@ static __always_inline int output_dns_event(struct __sk_buff *skb,
 
 	// Check network protocol.
 	// This only works with IPv4.
-	// For IPv6, gadget_socket_lookup() in pkg/gadgets/internal/socketenricher/bpf/sockets-map.h
+	// For IPv6, gadget_socket_lookup() in pkg/gadgets/socketenricher/bpf/sockets-map.h
 	// provides an example how to parse ip/ports on IPv6.
 	event->src.proto = event->dst.proto =
 		load_byte(skb, ETH_HLEN + offsetof(struct iphdr, protocol));
