+++ pkg/socketenricher/bpf/socket-enricher-helpers.h
