+++ pkg/socketenricher/bpf/sockets-iter.bpf.c
