+++ pkg/socketenricher/socketsiter_bpfel_x86.go
